cleanup unused imports. move state in Notes page down to NoteTable component
