+++ client/src/components/TextEditor/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { Editor } from "@tinymce/tinymce-react";
 import PropTypes from "prop-types";
